roles manage

+++ src/app/dashboard/management/roles/components/DeleteDialog.tsx
@@ -3,7 +3,6 @@
 import { useAtom, useAtomValue, useSetAtom } from 'jotai'
 import { Ban, Trash2 } from 'lucide-react'
 import { useState, useTransition } from 'react'
-
 import { deleteRole } from '@/actions/roles.actions'
 import { Button } from '@/components/Button'
 import {
@@ -14,6 +13,7 @@ import {
     DialogTitle,
 } from '@/components/Dialog'
 import { MessageError } from '@/components/Error'
+import { DEFAULT_ROLES } from '@/constants/client'
 import {
     openDeleteAtom,
     rolesAtom,
@@ -44,6 +44,17 @@ export function DeleteDialog() {
                 <form
                     action={() => {
                         startTransition(async () => {
+                            if (
+                                Object.values(DEFAULT_ROLES)
+                                    .map(r => r.toLowerCase())
+                                    .includes(entity.name.toLowerCase())
+                            ) {
+                                setMessage(
+                                    'El rol es administrado por el sistema, no se puede eliminar',
+                                )
+                                setTimeout(() => setMessage(''), 5_000)
+                                return
+                            }
                             const response = await deleteRole(entity.id)
                             if (response.status === 'error') {
                                 setMessage('Algo salio mal, intente de nuevo')
